+++ db/migrate/20191123144309_add_effect_to_recipes.rb
@@ -1,0 +1,6 @@
+class AddEffectToRecipes < ActiveRecord::Migration[5.2]
+  def change
+    add_column :recipes, :effect_name, :jsonb, default: { en: '', ru: '' }
+    add_column :recipes, :effect_links, :jsonb, default: { en: '', ru: '' }
+  end
+end